pyproject.toml
[project]
name = "voice-dataset-creation"
version = "0.1.0"
description = "Community-controlled voice data collection for language preservation and AI development"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20250625",
    "transformers>=5.5.0",
    "protobuf>=3.20.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "audiomentations>=0.33.0",
    "pandas>=3.0.2",
    "tqdm>=4.66.0",
    "soundfile>=0.12.1",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
